# openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: Render API
  description: REST API for the Render paint festival platform.
  version: "0.1.0"
  contact:
    name: Render
    url: https://renderltd.com
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.renderltd.com
    description: Production
...
components:
  schemas:
    Problem:
      description: RFC 7807 problem details object returned on all error responses.
      type: object
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
          example: "about:blank"
        title:
          type: string
          example: "Unprocessable Entity"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "email is required"
        instance:
          type: string
          format: uri
    User:
      type: object
      required: [id, email, is_admin, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          minLength: 8
          example: hunter2hunter
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    UnprocessableEntity:
      description: Validation error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
paths:
...
  /auth/signup:
    post:
      operationId: postAuthSignup
      summary: Create a new user account
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Email already registered.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /auth/login:
    post:
      operationId: postAuthLogin
      summary: Authenticate and receive a JWT
      description: Sets an HTTP-only session cookie and returns the token in the body for mobile clients.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      summary: Request password reset email
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "202":
          description: Reset email sent (returned regardless of whether email exists).
  /auth/reset-password:
    post:
      operationId: resetPassword
      summary: Reset password using token from email
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password updated.
        "400":
          description: Invalid or expired token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /me:
    get:
      operationId: getMe
      summary: Return the authenticated user
      tags: [auth]
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /images/presign:
    post:
      operationId: postImagesPresign
      summary: Request a presigned PUT URL for direct upload to object storage
      description: |
        Returns a URL the client uses to PUT image bytes directly to MinIO (dev) or S3 (prod).
        The API server is not in the upload data path. URL is valid for 15 minutes.
        Call POST /images/confirm after the PUT succeeds.
      tags: [images]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresignRequest"
      responses:
        "200":
          description: Presigned URL issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /images/confirm:
    post:
      operationId: postImagesConfirm
      summary: Confirm a completed upload and receive the CDN URL
      description: |
        Call after successfully PUTting bytes to the presigned URL.
        The server verifies the object exists in storage and returns a public CDN URL.
        Resource association is deferred to E5.
      tags: [images]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRequest"
      responses:
        "200":
          description: Upload confirmed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /profiles:
    post:
      operationId: postProfiles
      summary: Create artist profile
      description: Creates an artist profile for the authenticated user. One profile per user.
      tags: [artist]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileRequest"
      responses:
        "201":
          description: Profile created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /profiles/me:
    get:
      operationId: getProfileMe
      summary: Get own artist profile
      tags: [artist]
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Own profile (includes private fields like show_location setting).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchProfileMe
      summary: Update own artist profile
      tags: [artist]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /profiles/me/qr:
    get:
      operationId: getMyProfileQR
      summary: Get a branded QR code (PNG) for the authenticated artist's public profile
      tags: [profiles]
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Branded QR code encoding the artist's public profile URL.
          content:
            image/png:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /profiles/{profileID}:
    get:
      operationId: getProfile
      summary: Get public artist profile
      tags: [artist]
      parameters:
        - name: profileID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Public artist profile. location_label is omitted if artist has not enabled show_location.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
        "404":
          $ref: "#/components/responses/NotFound"
  /profiles/{profileID}/collections:
    get:
      operationId: listCollections
      summary: List collections for an artist profile
      tags: [artist]
      parameters:
        - name: profileID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Array of collections ordered by display_order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
  /profiles/{profileID}/festivals:
    get:
      operationId: listArtistFestivals
      summary: List an artist's public festival appearances
      description: >-
        Public. Returns the publicly-visible festivals (open or live) where this
        artist has an accepted application and/or an assigned spot, for the
        "Appearances" section on the public profile. map_slug is non-null only
        for live festivals (the public map renders for live festivals only).
      tags: [artist]
      parameters:
        - name: profileID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Array of festival appearances (empty array if none).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FestivalAppearance"
        "400":
          $ref: "#/components/responses/BadRequest"
  /collections:
    post:
      operationId: postCollections
      summary: Create a collection
      description: Creates a collection for the authenticated artist's profile. Requires an existing artist profile.
      tags: [artist]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionRequest"
      responses:
        "201":
          description: Collection created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /collections/order:
    put:
      operationId: reorderCollections
      summary: Reorder collections
      description: |
        Accepts an ordered array of collection IDs belonging to the authenticated artist.
        Sets display_order by position (0-indexed). Returns 204 on success.
      tags: [artist]
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderCollectionsRequest"
      responses:
        "204":
          description: Order saved.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /collections/{collectionID}:
    get:
      operationId: getCollection
      summary: Get a collection
      tags: [artist]
      parameters:
        - name: collectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchCollection
      summary: Update a collection
      tags: [artist]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: collectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCollectionRequest"
      responses:
        "200":
          description: Updated collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCollection
      summary: Delete a collection
      tags: [artist]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: collectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /collections/{collectionID}/images:
    get:
      operationId: listCollectionImages
      summary: List images in a collection
      tags: [artist]
      security: []
      parameters:
        - name: collectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of collection images ordered by display_order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollectionImage"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: attachCollectionImage
      summary: Attach an uploaded image to a collection
      description: |
        Attach an image that has already been confirmed via POST /images/confirm.
        Appends to the end of the collection's image list (display_order = current count).
      tags: [artist]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: collectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachImageRequest"
      responses:
        "201":
          description: Image attached.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionImage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /collections/{collectionID}/images/order:
    put:
      operationId: reorderCollectionImages
      summary: Reorder images in a collection
      description: |
        Accepts an ordered array of image IDs. Sets display_order by position (0-indexed).
        Returns the images in the new order.
      tags: [artist]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: collectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderImagesRequest"
      responses:
        "200":
          description: Images in new order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollectionImage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /collections/{collectionID}/images/{imageID}:
    delete:
      operationId: deleteCollectionImage
      summary: Remove an image from a collection
      tags: [artist]
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: collectionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: imageID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...